>>setting our brush color
====  our brush color =======

uncommented setting brush color event listener (37-39)
note: we won't be able to test this yet as we won't yet do the drawing in the canvas with the brush thingy

note2: the way we set our brush and eraser, is kinda the same thing.
The eraser is actually also a paintbrush that has the same color as the canvas
--that's why, we need to keep track if were painting with the brushColor or the eraser
-with that in mind, we've uncommented the isEraser global var (22)

made the currentColor var to have the same value as the brushColorBtn (39) (in our html, it's line 24)

/////// our brush color //////////

======= eraser =======
uncommented the eraser event lister (48-54)
- changed isEarser to true (50)
note: those lines (51 & 52) represents what were clicking. If it's black, it's selected (52), if it's white, it's not (51). It's the same thing that's happening in lines 60 and 61
-added isEraser to true since were working on an eraser (50)
- made the eraser to be the same color as canvas (54)
- made the eraser size to be 50px instead of the normal 10px (55)

uncommented the switchToBrush() (58-66)
- so again, depending on wheter you click on the eraser or the brush it will turn black indicating that it's selected (52 & 62), and the text too will change (53 & 61).
-The sizes also change, if it's an eraser, it's 50 (55) and if it's the brush it's 10 (65)
- notice that switchBrush is a function and not an event listener, that's because we'll use it more than once.

uncommeneted brushIcon.addEventListener (198-199)

trying it out:
(once again, drawing functionality not yet implemented at this stage.)
if you click on the brush icon, it turns black and says 'brush' on the top left

added switchToBrush() in our createCanvas(). We want that once we launch our app, the brush will be selected, ready to go

trying it out:
once loaded, the paint brush icon is black, indicating that it's selected.
if you click on the eraser, it turns black and says 'eraser' on the top left