>>dealing with the input range
(as of the moment, before this, the range in the brush doesn't work.
we want the no. 10 thingy to update as we move the slider.)

uncommented the setting brush size event listener (32-34)

made the currentSize to be the same size as the brushSlider.value (33)

uncommented displayBrushSize() (27-29) -- this is responsible to make the 10 thigy to be dynamic

added a conditional statement (28-32)

placed the displayBrushSize() in our brush slider event listener (38)

trying it out:
it does work
but,
problem:
when we switch from the eraser to brush, it doesn't reflect the change of defaulting back to 10

solution: adding code on line 71-72